Read double values correctly and store diskLoc as double

+++ tools/library/tbc/lddecodemetadata.h
@@ -201,7 +201,7 @@ public:
         DropOuts dropOuts;
         bool pad = false;
 
-        qint32 diskLoc = -1;
+        double diskLoc = -1;
         qint64 fileLoc = -1;
         qint32 decodeFaults = -1;
         qint32 efmTValues = -1;
